Moved GhostPathEditor into its own file in the Editor folder, so it is excluded when building for Android
