+++ Assets/Scripts/Editor/GhostPathEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d65677719dd1bbf44956f31c74bede85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
